Move HTML sanitisation to lib an add options param
