--- packages/frontend/amp/lib/sanitise-html.ts
@@ -1,3 +1,0 @@
-import sanitiszeHtml from 'sanitize-html';
-
-export const sanitise = (html: string): string => sanitiszeHtml(html);
