+++ packages/frontend/lib/sanitise-html.ts
@@ -1,0 +1,4 @@
+import sanitiszeHtml from 'sanitize-html';
+
+export const sanitise = (html: string, options: {} = {}): string =>
+    sanitiszeHtml(html, options);
